Move Hannah's cuts to gump_cuts.py
